Restructure Xray code (#930)
